backend/requirements.txt
# ──────────────────────────────────────────────
#  ETIP Backend — Python Dependencies
# ──────────────────────────────────────────────

# Core
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0

# Authentication
djangorestframework-simplejwt==5.3.1
django-otp==1.5.4
qrcode[pil]==8.0

# Database (psycopg3 — supports Python 3.14+)
psycopg[binary]>=3.2.10
psycopg[pool]>=3.2.10
django-environ==0.11.2

# Caching & Task Queue
redis==5.2.1
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
flower==2.0.1

# HTTP Client (for DISCO & gateway API calls)
httpx==0.28.1
tenacity==9.0.0

# Encryption
cryptography==44.0.0

# PDF generation (receipts) — cross-platform
xhtml2pdf==0.2.16

# Notifications
firebase-admin==6.6.0
twilio==9.4.5

# Storage
boto3==1.35.91
django-storages==1.14.4

# Utilities
python-decouple==3.8
Pillow>=12.0.0
phonenumbers==8.13.53
django-filter==24.3
drf-spectacular==0.27.2

# Security
django-ratelimit==4.1.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
faker==33.3.0
freezegun==1.5.1
httpretty==1.1.4

# Code quality
black==24.10.0
flake8==7.1.1
isort==5.13.2
